Fix all instances of 4244 and 4267 warnings in OV EP code (#23567)
### Description
Remove MSVC warnings 4244, 4267 from the list of disabled warnings in
cmake.
Fix the code that generates the warnings so that it no longer does.

### Motivation and Context
This makes onnxruntime_providers_openvino.dll pass BinSkim analysis.
Without this change BinSkim complains about the disabled warnings.

+++ onnxruntime/core/providers/openvino/backends/basic_backend.cc
@@ -129,7 +129,7 @@ BasicBackend::BasicBackend(std::unique_ptr<ONNX_NAMESPACE::ModelProto>& model_pr
   } catch (const char* msg) {
     ORT_THROW(msg);
   }
-  int num_infer_req = (global_context_.num_of_threads > 0) ? global_context_.num_of_threads : 1;
+  size_t num_infer_req = (global_context_.num_of_threads > 0) ? global_context_.num_of_threads : 1;
   inferRequestsQueue_ = std::unique_ptr<InferRequestsQueue>(new InferRequestsQueue(exe_network_, num_infer_req));
 }
 
@@ -339,7 +339,7 @@ void BasicBackend::EnableStreams() {
 void BasicBackend::SetNumThreads(ov::AnyMap& device_config) {
   // inference_num_threads is applicable only for the CPU device
   if (global_context_.device_type.find("CPU") != std::string::npos)
-    device_config.emplace(ov::inference_num_threads(global_context_.num_of_threads));
+    device_config.emplace(ov::inference_num_threads(static_cast<int>(global_context_.num_of_threads)));
 }
 
 // Starts an asynchronous inference request for data in slice indexed by batch_slice_idx on
